Fixed multiple includes , didn't actually need to use Character struct in platformer_combat.cpp
